+++ frontend/packages/data-portal/app/components/BrowseData/DatasetTable.tsx
@@ -14,6 +14,7 @@ import { Link } from 'app/components/Link'
 import { CellHeader, PageTable, TableCell } from 'app/components/Table'
 import { EMPIAR_ID, EMPIAR_URL } from 'app/constants/external-dbs'
 import { ANNOTATED_OBJECTS_MAX, MAX_PER_PAGE } from 'app/constants/pagination'
+import { DatasetTableWidths } from 'app/constants/table'
 import { Dataset, useDatasets } from 'app/hooks/useDatasets'
 import { useI18n } from 'app/hooks/useI18n'
 import { useIsLoading } from 'app/hooks/useIsLoading'
@@ -55,9 +56,37 @@ export function DatasetTable() {
 
     try {
       return [
+        columnHelper.accessor('key_photo_thumbnail_url', {
+          header: () => <p />,
+
+          cell({ row: { original: dataset } }) {
+            const previousUrl = `${location.pathname}${location.search}`
+            const datasetUrl = `/datasets/${
+              dataset.id
+            }?prev=${encodeURIComponent(previousUrl)}`
+
+            return (
+              <TableCell
+                renderLoadingSkeleton={false}
+                width={DatasetTableWidths.photo}
+              >
+                <Link to={datasetUrl} className="max-w-[134px]">
+                  <KeyPhoto
+                    className="max-w-[134px]"
+                    title={dataset.title}
+                    src={dataset.key_photo_thumbnail_url ?? undefined}
+                    loading={isLoadingDebounced}
+                  />
+                </Link>
+              </TableCell>
+            )
+          },
+        }),
+
         columnHelper.accessor('id', {
           header: () => (
             <CellHeader
+              showSort
               active={datasetSort !== undefined}
               direction={datasetSort}
               onClick={(event) => {
@@ -75,10 +104,12 @@ export function DatasetTable() {
 
                 setSearchParams(nextParams)
               }}
+              width={DatasetTableWidths.id}
             >
-              {t('dataset')}
+              {t('datasetName')}
             </CellHeader>
           ),
+
           cell({ row: { original: dataset } }) {
             const previousUrl = `${location.pathname}${location.search}`
             const datasetUrl = `/datasets/${
@@ -89,17 +120,8 @@ export function DatasetTable() {
               <TableCell
                 className="flex px-sds-s py-sds-l gap-sds-m"
                 renderLoadingSkeleton={false}
-                minWidth={450}
-                maxWidth={800}
+                width={DatasetTableWidths.id}
               >
-                <Link to={datasetUrl} className="flex-shrink-0 w-[134px]">
-                  <KeyPhoto
-                    title={dataset.title}
-                    src={dataset.key_photo_thumbnail_url ?? undefined}
-                    loading={isLoadingDebounced}
-                  />
-                </Link>
-
                 <div className="flex flex-col flex-auto gap-sds-xxxs min-h-[100px]">
                   <p className="text-sm font-semibold text-sds-primary-400">
                     {isLoadingDebounced ? (
@@ -113,7 +135,7 @@ export function DatasetTable() {
                     {isLoadingDebounced ? (
                       <Skeleton className="max-w-[120px]" variant="text" />
                     ) : (
-                      `${t('portalId')}: ${dataset.id}`
+                      `${t('datasetId')}: ${dataset.id}`
                     )}
                   </p>
 
@@ -161,13 +183,18 @@ export function DatasetTable() {
         }),
 
         columnHelper.accessor('related_database_entries', {
-          header: t('empiarID'),
+          header: () => (
+            <CellHeader width={DatasetTableWidths.empiarId}>
+              {t('empiarID')}
+            </CellHeader>
+          ),
+
           cell({ getValue }) {
             const empiarIDMatch = EMPIAR_ID.exec(getValue() ?? '')
             const empiarID = empiarIDMatch?.[1]
 
             return (
-              <TableCell minWidth={120} maxWidth={130}>
+              <TableCell width={DatasetTableWidths.empiarId}>
                 {empiarID ? (
                   <Link
                     className="text-sds-primary-500 inline"
@@ -184,12 +211,16 @@ export function DatasetTable() {
         }),
 
         columnHelper.accessor('organism_name', {
-          header: t('organismName'),
+          header: () => (
+            <CellHeader width={DatasetTableWidths.organismName}>
+              {t('organismName')}
+            </CellHeader>
+          ),
+
           cell: ({ getValue }) => (
             <TableCell
               primaryText={getValue() ?? '--'}
-              minWidth={100}
-              maxWidth={400}
+              width={DatasetTableWidths.organismName}
             />
           ),
         }),
@@ -198,28 +229,35 @@ export function DatasetTable() {
           (dataset) => dataset.runs_aggregate.aggregate?.count,
           {
             id: 'runs',
+
             header: () => (
               <CellHeader
-                hideSortIcon
                 tooltip={<I18n i18nKey="runsTooltip" />}
                 arrowPadding={{ right: 270 }}
+                width={DatasetTableWidths.runs}
               >
                 {t('runs')}
               </CellHeader>
             ),
+
             cell: ({ getValue }) => (
               <TableCell
                 primaryText={String(getValue() ?? 0)}
-                minWidth={70}
-                maxWidth={100}
+                width={DatasetTableWidths.runs}
               />
             ),
           },
         ),
 
         columnHelper.accessor((dataset) => dataset.runs, {
-          // TODO: fix overflow for this header
-          header: t('annotatedObjects'),
+          id: 'annotatedObjects',
+
+          header: () => (
+            <CellHeader width={DatasetTableWidths.annotatedObjects}>
+              {t('annotatedObjects')}
+            </CellHeader>
+          ),
+
           cell({ getValue }) {
             const runs = getValue()
             const annotatedObjects = Array.from(
@@ -236,9 +274,7 @@ export function DatasetTable() {
 
             return (
               <TableCell
-                minWidth={120}
-                maxWidth={400}
-                className="w-[15%]"
+                width={DatasetTableWidths.annotatedObjects}
                 renderLoadingSkeleton={() => (
                   <div className="flex flex-col gap-2">
                     {range(0, ANNOTATED_OBJECTS_MAX).map((val) => (