+++ frontend/packages/data-portal/app/components/Run/RunHeader.tsx
@@ -1,12 +1,13 @@
 import { Button, Icon } from '@czi-sds/components'
-import { sum } from 'lodash-es'
 
 import { I18n } from 'app/components/I18n'
 import { KeyPhoto } from 'app/components/KeyPhoto'
 import { Link } from 'app/components/Link'
 import { PageHeader } from 'app/components/PageHeader'
+import { PageHeaderSubtitle } from 'app/components/PageHeaderSubtitle'
 import { MetadataTable } from 'app/components/Table'
 import { TiltSeriesQualityScoreBadge } from 'app/components/TiltSeriesQualityScoreBadge'
+import { ViewTomogramButton } from 'app/components/ViewTomogramButton'
 import { useDownloadModalQueryParamState } from 'app/hooks/useDownloadModalQueryParamState'
 import { useI18n } from 'app/hooks/useI18n'
 import {
@@ -16,8 +17,6 @@ import {
 import { useRunById } from 'app/hooks/useRunById'
 import { i18n } from 'app/i18n'
 import { getTiltRangeLabel } from 'app/utils/tiltSeries'
-
-import { ViewTomogramButton } from '../ViewTomogramButton'
 
 export function RunHeader() {
   const { run } = useRunById()
@@ -69,42 +68,12 @@ export function RunHeader() {
       }
       backToResultsLabel={run.dataset.title}
       lastModifiedDate="2023-12-16"
-      metadata={[
-        // TODO fetch frames from API
-        { key: i18n.frames, value: i18n.nFiles(0) },
-
-        {
-          key: i18n.tiltSeries,
-          value: i18n.nFiles(run.tiltseries_aggregate.aggregate?.count ?? 0),
-        },
-
-        {
-          key: i18n.tomograms,
-          value: i18n.nFiles(
-            sum(
-              run.tomogram_stats.flatMap(
-                (stats) => stats.tomograms_aggregate.aggregate?.count ?? 0,
-              ),
-            ),
-          ),
-        },
-
-        {
-          key: i18n.annotations,
-          value: i18n.nFiles(
-            sum(
-              run.tomogram_stats.flatMap(
-                (stats) => stats.annotations_aggregate.aggregate?.count ?? 0,
-              ),
-            ),
-          ),
-        },
-      ]}
+      metadata={[{ key: t('runId'), value: `${run.id}` }]}
       onMoreInfoClick={() => toggleDrawer(MetadataDrawerId.Run)}
       title={run.name}
       renderHeader={({ moreInfo }) => (
-        <div className="flex gap-sds-xxl p-sds-xl border-t-[3px] border-sds-gray-200">
-          <div className="max-w-[300px] max-h-[213px] grow overflow-clip rounded-sds-m flex-shrink-0 flex items-center">
+        <div className="flex gap-sds-xxl p-sds-xl">
+          <div className="max-w-[465px] max-h-[330px] grow overflow-clip rounded-sds-m flex-shrink-0 flex items-center">
             {keyPhotoURL ? (
               <Link to={keyPhotoURL}>
                 <KeyPhoto title={run.name} src={keyPhotoURL} />
@@ -114,11 +83,15 @@ export function RunHeader() {
             )}
           </div>
 
-          <div className="flex flex-col gap-sds-xl flex-auto">
+          <div className="flex flex-col gap-sds-xl flex-auto pt-sds-l">
+            <PageHeaderSubtitle>{t('runOverview')}</PageHeaderSubtitle>
+
             <div className="flex gap-sds-xxl flex-col lg:flex-row">
               <MetadataTable
                 title={i18n.tiltSeries}
-                tableCellLabelProps={{ maxWidth: 100, minWidth: 100 }}
+                tableCellLabelProps={{
+                  width: { min: 100, max: 100 },
+                }}
                 data={[
                   {
                     labelTooltip: <I18n i18nKey="tiltSeriesTooltip" />,
@@ -159,7 +132,7 @@ export function RunHeader() {
 
               <MetadataTable
                 title={i18n.tomogram}
-                tableCellLabelProps={{ maxWidth: 180, minWidth: 100 }}
+                tableCellLabelProps={{ width: { min: 100, max: 180 } }}
                 data={[
                   {
                     label: i18n.resolutionsAvailable,