+++ frontend/packages/data-portal/app/context/DownloadModal.context.ts
@@ -9,9 +9,10 @@ export type TomogramResolution =
   GetRunByIdQuery['runs'][number]['tomogram_stats'][number]['tomogram_resolutions'][number]
 
 export interface DownloadModalContextValue {
+  activeAnnotation?: Annotation | null
+  activeTomogramResolution?: TomogramResolution | null
   allTomogramProcessing?: string[]
   allTomogramResolutions?: TomogramResolution[]
-  allAnnotations?: Map<number, Annotation>
   datasetId?: number
   datasetTitle?: string
   fileSize?: number
@@ -19,9 +20,7 @@ export interface DownloadModalContextValue {
   objectName?: string
   runId?: number
   runName?: string
-  s3DatasetPrefix?: string
-  s3TomogramPrefix?: string
-  s3TomogramVoxelPrefix?: string
+  s3Path?: string
   tomogramId?: number
   tomogramVoxelId?: number
   type: DownloadModalType