+++ frontend/packages/data-portal/app/hooks/useRunById.ts
@@ -5,14 +5,8 @@ import { GetRunByIdQuery } from 'app/__generated__/graphql'
 
 export function useRunById() {
   const {
-    data: {
-      runs: [run],
-    },
-    fileSizeMap,
-  } = useTypedLoaderData<{
-    data: GetRunByIdQuery
-    fileSizeMap: Record<string, number>
-  }>()
+    runs: [run],
+  } = useTypedLoaderData<GetRunByIdQuery>()
 
   const objectNames = useMemo(
     () =>
@@ -58,7 +52,6 @@ export function useRunById() {
 
   return {
     run,
-    fileSizeMap,
     objectNames,
     objectShapeTypes,
     annotationSoftwares,